Level Transition fix
Changed the system from using AController* to FString. I'm not too happy with this change but it is necessary to have it working with level transitions as it is not possible from the controller or player state to ensure that the player is the same as before or identify the correct player for each quest key. Yes, I could've used the UniqueNetID however I decided not to as the string can be obtained from the NetID and it can be used for non-Players.

Names have been changed accordingly